snapcraft.yaml: use snapcraft's hook generation instead of command-chain
Using command-chain here is a bit side-stepping snapcraft and using the wrapper
anyways, but we don't need the wrapper for our Go app anyways so we can just get
rid of it in a follow-up commit and have a cleaner snap.

This requires staging the hooks dir into $SNAPCRAFT_PRIME/snap using the dump
plugin.

Signed-off-by: Ian Johnson <ian.johnson@canonical.com>